im done

+++ Refresh.GameServer/Types/Telemetry/Json/Events/TelemetryNetStats.cs
@@ -4,9 +4,9 @@ public class TelemetryNetStats
 {
     [JsonProperty("IsLocal")] public bool IsLocal { get; set; }
     [JsonProperty("Player")] public string Player { get; set; }
-    [JsonProperty("AvailBW")] public int AvailableBandwidth { get; set; }
-    [JsonProperty("AvailRNPBW")] public int AvailableRnpBandwidth { get; set; }
-    [JsonProperty("AvailGameBW")] public double AvailableGameBandwidth { get; set; }
-    [JsonProperty("RecentTotBWUsed")] public int RecentTotalBandwidthUsed { get; set; }
-    [JsonProperty("TimeBtwnSends")] public double TimeBetweenSends { get; set; }
+    [JsonProperty("AvailBW")] public uint AvailableBandwidth { get; set; }
+    [JsonProperty("AvailRNPBW")] public uint AvailableRnpBandwidth { get; set; }
+    [JsonProperty("AvailGameBW")] public float AvailableGameBandwidth { get; set; }
+    [JsonProperty("RecentTotBWUsed")] public uint RecentTotalBandwidthUsed { get; set; }
+    [JsonProperty("TimeBtwnSends")] public float TimeBetweenSends { get; set; }
 }
